refactor: 로그인 로직 및 전역 사용법 수정
