+++ app/todoList/page.tsx
@@ -6,9 +6,9 @@ import { useInView } from 'react-intersection-observer';
 import { useInfiniteQuery, useQueryClient } from '@tanstack/react-query';
 
 import { readTodoList } from '@/api/todo';
+import { useInfoStore } from '@/provider/store-provider';
 import PlusIcon from '@/public/icon/plus-border-none.svg';
 import TodoListIcon from '@/public/icon/todo-list-black.svg';
-import { useInfoStore } from '@/stores/infoStore';
 
 import EmptyMessage from './_components/EmptyMessage';
 import PriorityFilter from './_components/PriorityFilter';
@@ -20,7 +20,7 @@ import type { QueryClient } from '@tanstack/react-query';
 
 export default function TodoListPage() {
   const queryClient: QueryClient = useQueryClient();
-  const { id: userId } = useInfoStore();
+  const { id: userId } = useInfoStore((state) => state);
   const [filter, setFilter] = useState<FilterType>('All');
   const [priority, setPriority] = useState<'all' | number>('all');
